Reorganize prediction models hierarchically
