+++ src/tempor/plugins/_plugin_loader.py
@@ -1,14 +1,13 @@
 import tempor.plugins.core as plugins
 
-from . import classification, preprocessing, regression, time_to_event, treatments
+from . import prediction, preprocessing, time_to_event, treatments
 
 plugin_loader = plugins.PluginLoader()
 
 
 __all__ = [
-    "classification",
+    "prediction",
     "preprocessing",
-    "regression",
     "time_to_event",
     "treatments",
 ]
