+++ src/tempor/plugins/prediction/__init__.py
@@ -1,0 +1,6 @@
+from . import one_off, temporal
+
+__all__ = [
+    "one_off",
+    "temporal",
+]
