+++ src/tempor/plugins/prediction/one_off/__init__.py
@@ -1,0 +1,6 @@
+from . import classification, regression
+
+__all__ = [
+    "classification",
+    "regression",
+]
