+++ src/tempor/plugins/prediction/temporal/__init__.py
@@ -1,0 +1,6 @@
+from . import classification, regression
+
+__all__ = [
+    "classification",
+    "regression",
+]
